feat: Reorganize UI structure and centralize CSS variables
- Created organized folder structure for UI components
- Centralized all CSS variables in Global.css for consistency
- Compressed and updated background animation

- Created `components/ui/icons/` directory for all image assets
- Moved all CSS files from `components/ui/` to `components/styles/`
- Updated all import paths across the codebase

- Added comprehensive CSS custom properties in `Global.css`:
  - Color palette (primary, secondary, text, background, borders)
  - Spacing system (xs to 4xl)
  - Typography scales and font families
  - Visual effects (shadows, blur, transitions)
  - Z-index management
- Updated multiple CSS files to use centralized variables:
  - `AppLayout.css`, `HomePage.css`, `HomeConnectedPage.css`
  - `BottomNavigation.css`, `ChatPage.css`

- Organized icons into logical subfolder structure
- Compressed background animation for better performance
- Updated all component imports to reflect new paths

- Eliminated duplicate CSS values across files
- Improved maintainability with consistent design tokens
- Enhanced theme management capabilities
