+++ SofIA/extension/components/layout/AppLayout.tsx
@@ -1,6 +1,7 @@
 import { useStorage } from "@plasmohq/storage/hook"
 import SplineBackground from '../Splinebackground'
 import { useRouter } from './RouterProvider'
+import '../styles/Global.css'
 import '../styles/AppLayout.css'
 
 interface AppLayoutProps {
