+++ SofIA/extension/components/pages/SearchResultPage.tsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from 'react'
 import { useRouter } from '../layout/RouterProvider'
 import { useMCPClient, type SearchResult } from '../../hooks/useMCPClient'
-import SendIcon from '../ui/quick_action/Selected=send.svg'
-import SendHoverIcon from '../ui/quick_action/Selected=send hover.svg'
-import VoteIcon from '../ui/quick_action/Selected=vote.svg'
-import VoteHoverIcon from '../ui/quick_action/Selected=vote hover.svg'
+import SendIcon from '../ui/icons/quick_action/Selected=send.svg'
+import SendHoverIcon from '../ui/icons/quick_action/Selected=send hover.svg'
+import VoteIcon from '../ui/icons/quick_action/Selected=vote.svg'
+import VoteHoverIcon from '../ui/icons/quick_action/Selected=vote hover.svg'
 import PersonIcon from '../../assets/Icon=person.svg'
 import VoteIconSvg from '../../assets/Icon=vote.svg'
 import '../styles/Global.css'
