+++ SofIA/extension/components/styles/AppLayout.css
@@ -2,7 +2,7 @@
   position: relative;
   width: 100%;
   min-height: 100vh;
-  font-family: 'Montserrat', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  font-family: var(--font-family-primary);
 }
 
 .app-overlay {
@@ -11,13 +11,13 @@
   left: 0;
   right: 0;
   bottom: 0;
-  background-color: rgba(14, 14, 14, 0.30);
-  z-index: 1;
+  background-color: var(--color-bg-dark);
+  z-index: var(--z-base);
 }
 
 .app-content {
   position: relative;
   z-index: 2;
-  min-height: calc(100vh - 60px);
-  padding-bottom: 60px;
+  min-height: calc(100vh - var(--size-nav-height));
+  padding-bottom: var(--size-nav-height);
 }
