+++ SofIA/extension/components/styles/AtomCreationModal.css
