+++ SofIA/extension/components/styles/Global.css
@@ -1,62 +1,166 @@
 /* Styles globaux et communs à toute l'application */
 
+:root {
+  /* Couleurs principales */
+  --color-primary: #C7866C;
+  --color-primary-dark: #B8956A;
+  --color-primary-light: #D4A574;
+  --color-secondary: #DB6B3E;
+  --color-accent: #A6AF6B;
+  --color-hover: #FFB3C4;
+  --color-hover-light: #FFE078;
+  
+  /* Couleurs de texte */
+  --color-text-primary: #F2DED6;
+  --color-text-secondary: #FBF7F5;
+  --color-text-dark: #050505;
+  --color-text-muted: rgba(255, 255, 255, 0.8);
+  --color-text-very-muted: rgba(255, 255, 255, 0.6);
+  --color-text-placeholder: rgba(255, 255, 255, 0.5);
+  
+  /* Couleurs de fond */
+  --color-bg-glass: rgba(255, 255, 255, 0.02);
+  --color-bg-glass-light: rgba(255, 255, 255, 0.05);
+  --color-bg-glass-lighter: rgba(255, 255, 255, 0.1);
+  --color-bg-dark: rgba(14, 14, 14, 0.30);
+  --color-bg-darker: rgba(10, 10, 10, 0.7);
+  --color-bg-black: rgba(0, 0, 0, 0.8);
+  --color-bg-overlay: rgba(26, 26, 26, 0.95);
+  
+  /* Bordures */
+  --color-border-glass: rgba(255, 255, 255, 0.08);
+  --color-border-light: rgba(255, 255, 255, 0.1);
+  --color-border-lighter: rgba(255, 255, 255, 0.125);
+  --color-border-very-light: rgba(255, 255, 255, 0.05);
+  --color-border-muted: rgba(242, 222, 214, 0.2);
+  
+  /* Couleurs d'état */
+  --color-success: #22c55e;
+  --color-error: #ef4444;
+  --color-warning: #f59e0b;
+  --color-info: #60a5fa;
+  
+  /* Effets visuels */
+  --backdrop-blur: blur(10px);
+  --backdrop-blur-light: blur(5px);
+  --backdrop-blur-heavy: blur(20px);
+  --backdrop-saturate: saturate(100%);
+  
+  /* Ombres */
+  --shadow-glass: 0 4px 24px rgba(0, 0, 0, 0.2);
+  --shadow-heavy: 0 8px 32px rgba(0, 0, 0, 0.3);
+  --shadow-light: 0 2px 6px rgba(0, 0, 0, 0.2);
+  --shadow-primary: 0 4px 16px rgba(219, 107, 62, 0.3);
+  --shadow-button: 0 4px 12px rgba(212, 165, 116, 0.3);
+  --shadow-button-hover: 0 6px 16px rgba(212, 165, 116, 0.4);
+  --shadow-inset: inset 0 1px 0 rgba(255, 255, 255, 0.2);
+  --shadow-inset-light: inset 0 1px 0 rgba(255, 255, 255, 0.1);
+  
+  /* Transitions */
+  --transition-fast: all 0.2s ease;
+  --transition-normal: all 0.3s ease;
+  --transition-slow: all 0.5s ease;
+  
+  /* Espacements */
+  --spacing-xs: 4px;
+  --spacing-sm: 8px;
+  --spacing-md: 12px;
+  --spacing-lg: 16px;
+  --spacing-xl: 20px;
+  --spacing-2xl: 24px;
+  --spacing-3xl: 30px;
+  --spacing-4xl: 40px;
+  
+  /* Border radius */
+  --radius-sm: 4px;
+  --radius-md: 8px;
+  --radius-lg: 12px;
+  --radius-xl: 20px;
+  --radius-full: 50%;
+  
+  /* Tailles */
+  --size-icon-sm: 16px;
+  --size-icon-md: 20px;
+  --size-icon-lg: 24px;
+  --size-button-height: 60px;
+  --size-nav-height: 60px;
+  
+  /* Typographie */
+  --font-family-primary: 'Montserrat', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  --font-family-display: 'Gotu', cursive;
+  --font-size-xs: 10px;
+  --font-size-sm: 12px;
+  --font-size-base: 14px;
+  --font-size-lg: 16px;
+  --font-size-xl: 18px;
+  --font-size-2xl: 24px;
+  --font-size-3xl: 28px;
+  
+  /* Z-index */
+  --z-base: 1;
+  --z-nav: 3;
+  --z-overlay: 5;
+  --z-header: 100;
+  --z-modal: 999;
+  --z-tooltip: 9999;
+}
+
 /* Container principal pour les pages */
 .page {
-  padding: 20px;
-  background-color: rgba(255, 255, 255, 0.02);
-  margin: 10px;
-  border-radius: 20px;
-  backdrop-filter: blur(5px) saturate(100%);
-  -webkit-backdrop-filter: blur(5px) saturate(100%);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  box-shadow: 0 4px 24px rgba(0, 0, 0, 0.2);
-  transition: all 0.3s ease;
+  padding: var(--spacing-xl);
+  background-color: var(--color-bg-glass);
+  margin: var(--spacing-sm);
+  border-radius: var(--radius-xl);
+  backdrop-filter: var(--backdrop-blur-light) var(--backdrop-saturate);
+  -webkit-backdrop-filter: var(--backdrop-blur-light) var(--backdrop-saturate);
+  border: 1px solid var(--color-border-glass);
+  box-shadow: var(--shadow-glass);
+  transition: var(--transition-normal);
 }
 
 /* Bouton retour standard */
 .back-button {
   background-color: transparent;
   border: none;
-  color: #F2DED6;
-  font-size: 12px;
+  color: var(--color-text-primary);
+  font-size: var(--font-size-sm);
   cursor: pointer;
-  margin-bottom: 20px;
-  padding: 8px 16px;
-  transition: all 0.3s ease;
+  margin-bottom: var(--spacing-xl);
+  padding: var(--spacing-sm) var(--spacing-lg);
+  transition: var(--transition-normal);
   display: flex;
   justify-content: flex-start;
-  
 }
 
 /* Titre de section */
 .section-title {
-  font-family: 'Gotu', cursive;
-  font-size: 24px;
+  font-family: var(--font-family-display);
+  font-size: var(--font-size-2xl);
   font-weight: 600;
-  color: #050505;
-  margin-bottom: 15px;
+  color: var(--color-text-dark);
+  margin-bottom: var(--spacing-lg);
 }
 
 /* Contenu de page */
 .page-content {
-  margin-top: 20px;
+  margin-top: var(--spacing-xl);
 }
 
 /* État vide */
 .empty-state {
   text-align: center;
-  color: #F2DED6;
+  color: var(--color-text-primary);
   font-style: italic;
-  padding: 30px;
+  padding: var(--spacing-3xl);
 }
 
 /* État d'erreur */
 .error-state {
   text-align: center;
-  color: #ef4444;
+  color: var(--color-error);
   font-style: italic;
-  padding: 20px;
+  padding: var(--spacing-xl);
   background-color: rgba(239, 68, 68, 0.1);
-  border-radius: 12px;
+  border-radius: var(--radius-lg);
   border: 1px solid rgba(239, 68, 68, 0.2);
 }
