+++ SofIA/extension/components/ui/QuickActionButton.tsx
@@ -1,14 +1,14 @@
 import { useState } from 'react'
-import AddIcon from './quick_action/Selected=add.svg'
-import AddHoverIcon from './quick_action/Selected=add hover.svg'
-import RemoveIcon from './quick_action/Selected=remove.svg'
-import RemoveHoverIcon from './quick_action/Selected=remove hover.svg'
-import SendIcon from './quick_action/Selected=send.svg'
-import SendHoverIcon from './quick_action/Selected=send hover.svg'
-import ViewIcon from './quick_action/Selected=view.svg'
-import ViewHoverIcon from './quick_action/Selected=view hover.svg'
-import VoteIcon from './quick_action/Selected=vote.svg'
-import VoteHoverIcon from './quick_action/Selected=vote hover.svg'
+import AddIcon from './icons/quick_action/Selected=add.svg'
+import AddHoverIcon from './icons/quick_action/Selected=add hover.svg'
+import RemoveIcon from './icons/quick_action/Selected=remove.svg'
+import RemoveHoverIcon from './icons/quick_action/Selected=remove hover.svg'
+import SendIcon from './icons/quick_action/Selected=send.svg'
+import SendHoverIcon from './icons/quick_action/Selected=send hover.svg'
+import ViewIcon from './icons/quick_action/Selected=view.svg'
+import ViewHoverIcon from './icons/quick_action/Selected=view hover.svg'
+import VoteIcon from './icons/quick_action/Selected=vote.svg'
+import VoteHoverIcon from './icons/quick_action/Selected=vote hover.svg'
 
 interface QuickActionButtonProps {
   action: 'add' | 'remove' | 'send' | 'view' | 'vote' | 'scan'
